重新构建结构
