+++ Attemper_Algorithm/Algorithm.cpp
@@ -1,33 +1,33 @@
 #include "Algorithm.h"
 
-bool Algorithm::FCFSAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::FCFSAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::SJFAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::SJFAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::HRRNAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::HRRNAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::SRTAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::SRTAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::SPFAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::SPFAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::RRAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::RRAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::nonPreemptivePrioritySchedulingAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::nonPreemptivePrioritySchedulingAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
 
-bool Algorithm::preemptivePrioritySchedulingAlgorithm(vector<Job> jobs, string fileName) {
-    return false;
-}
+//bool Algorithm::preemptivePrioritySchedulingAlgorithm(vector<Job> jobs, string fileName) {
+//    return false;
+//}
