+++ Attemper_Algorithm/Algorithm_FCFS.cpp
@@ -1,0 +1,11 @@
+#include "Algorithm.h"
+
+/**
+ * 先来线服务算法
+ * @param jobs 按找到达时间顺序排好顺序的作业集
+ * @param fileName 输出结果的文件名
+ * @return 成功返回true，否则返回false
+ */
+bool Algorithm::FCFSAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
