+++ Attemper_Algorithm/Algorithm_HRRN.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::HRRNAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
