+++ Attemper_Algorithm/Algorithm_RR.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::RRAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
