+++ Attemper_Algorithm/Algorithm_SJF.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::SJFAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
