+++ Attemper_Algorithm/Algorithm_SPF.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::SPFAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
