+++ Attemper_Algorithm/Algorithm_SRT.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::SRTAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
