+++ Attemper_Algorithm/Algorithm_nonPreemptivePriority.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::nonPreemptivePrioritySchedulingAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
