+++ Attemper_Algorithm/Algorithm_preemptivePriorityScheduling.cpp
@@ -1,0 +1,5 @@
+#include "Algorithm.h"
+
+bool Algorithm::preemptivePrioritySchedulingAlgorithm(vector<Job> jobs, string fileName) {
+    return false;
+}
