+++ Attemper_Algorithm/old/FormatInput.cpp
