+++ Attemper_Algorithm/old/FormatInput.h
@@ -1,7 +1,7 @@
 #ifndef ATTEMPER_ALGORITHM_FORMATINPUT_H
 #define ATTEMPER_ALGORITHM_FORMATINPUT_H
 
-#include "BASE.h"
+#include "../BASE.h"
 
 /**
  * 将输入的字符串转换为Job类型的结构体
