+++ Attemper_Algorithm/old/HRRN.cpp
